kiev: Remove unused media_codecs
kiev is SoC ID 459 with vidc SKU version 1, which uses the v3 configs
according to init.qcom.early_boot.sh.

Copy the v3 configs over the default ones, since we lack the qcom
framework changes to load the custom configs. This is no longer an
issue on newer platforms such as sm8350 which has switched to using
the AOSP SKU way to load different media configs.

Change-Id: Ib249132368f6c660e615cb53d4f37ffb71745e82 | AOSP